:lipstick: Update import statements in splitters
